cash flow component
